Build refactoring
# Conflicts:
#	src/html/build-plugins/feature-blocks-plugin.js
#	src/html/dev-plugins/dev-mock-plugin.js
